Conflicts:
	drivers/usb/musb/musb_gadget.c

+++ arch/arm/mach-omap2/omap_tps6236x.c
@@ -416,7 +416,7 @@ out:
 	return r;
 }
 
-int __init omap_tps6236x_update(char *name, u16 old_chip_id, u16 new_chip_id)
+int __init omap_tps6236x_update(char *name, u32 old_chip_id, u32 new_chip_id)
 {
 	return omap_pmic_update(omap_tps_map, name, old_chip_id, new_chip_id);
 }
